requirements.txt
numpy==1.26.4
pandas==2.2.2
opencv-python==4.10.0.84
scikit-learn==1.5.1
python-dotenv==1.0.1
Pillow==10.4.0
torch>=2.2.0
transformers==4.45.2
accelerate==0.34.2
bitsandbytes==0.43.3
sentencepiece==0.2.0
streamlit==1.38.0
plotly==5.24.1
tqdm>=4.66.0
reportlab==4.2.2
pydantic==2.8.2
openai==1.51.2
